Return the exact fractional utcoffset() instead of rounding it
Review on the previous commit (truncate instead of floor) showed
truncation is not actually a fix: for a :30 zone like India it is
30 minutes off, same as floor, just in the other direction, and for
a :45 zone (Nepal +5:45, Chatham/Eucla +8:45) truncation is 45
minutes off versus floor's 15 -- strictly worse, not better.

The rounding itself was never needed. utcoffset()'s only consumer,
adjust_timestamp(), does plain float arithmetic, and the value's
only other use is passed through the event dict in
events/dispatcher.py and read back via itemgetter in
events/receiver.py -- nothing requires an int. Returning
time.timezone/altzone / 3600 directly (no floor, no truncation) is
exact for every zone, fractional or whole-hour alike.

adjust_timestamp()'s offset parameter is widened from int to float
to match, since it now receives fractional values.

Tests updated: the India case now asserts the exact -5.5 instead of
pinning the previously-"fixed" but still-wrong -5, and a new test
covers a :45 zone (Nepal, -5.75), which is where floor and
truncation actually diverge in how wrong they are.

+++ celery/utils/time.py
@@ -453,21 +453,20 @@ def utcoffset(
         time: ModuleType = _time,
         localtime: Callable[..., _time.struct_time] = _time.localtime) -> float:
     """Return the current offset to UTC in hours."""
-    # Truncate toward zero, don't floor-divide: `time.timezone`/`time.altzone`
-    # are seconds *west* of UTC, so for a fractional-hour, east-of-UTC zone
-    # (e.g. India/Sri Lanka/Iran at UTC+5:30, seconds = -19800) a floor
-    # division rounds the magnitude *up* (-19800 // 3600 == -6), while the
-    # same fractional offset on the west-of-UTC side is unaffected (floor and
-    # truncation agree for positive dividends). That asymmetry skews
-    # adjust_timestamp() by up to an hour for those locales' events.
-    # LocalTimezone.__repr__ above already truncates the equivalent
-    # seconds-to-hours conversion; this matches it instead of floor-dividing.
+    # Use true division, not floor division: `time.timezone`/`time.altzone`
+    # are seconds *west* of UTC, and for a fractional-hour zone (e.g.
+    # India/Sri Lanka at UTC+5:30, seconds = -19800) floor-dividing rounds
+    # the magnitude *up* (-19800 // 3600 == -6) while the same fractional
+    # offset on the west-of-UTC side is unaffected (floor and truncation
+    # agree for positive dividends). Either kind of rounding is wrong for
+    # some fractional-hour zone or other -- `adjust_timestamp()` below is
+    # pure arithmetic on this value, so keep it exact instead.
     if localtime().tm_isdst:
-        return int(time.altzone / 3600)
-    return int(time.timezone / 3600)
+        return time.altzone / 3600
+    return time.timezone / 3600
 
 
-def adjust_timestamp(ts: float, offset: int,
+def adjust_timestamp(ts: float, offset: float,
                      here: Callable[..., float] = utcoffset) -> float:
     """Adjust timestamp based on provided utcoffset."""
     return ts - (offset - here()) * 3600
